requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
addict=2.4.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.05.30=h06a4308_0
certifi=2023.5.7=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
chardet=5.1.0=pypi_0
charset-normalizer=3.1.0=pypi_0
chumpy=0.70=pypi_0
click=8.1.3=pypi_0
colorama=0.4.6=pypi_0
comm=0.1.2=py310h06a4308_0
contourpy=1.1.0=pypi_0
cryptography=39.0.1=py310h9ce1e76_2
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
cython=3.0.0b3=pypi_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.2.2=h20bf706_0
fonttools=4.40.0=pypi_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
idna=3.4=py310h06a4308_0
importlib-metadata=6.7.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.19.2=py310h2f386ee_0
ipython=8.12.0=py310h06a4308_0
jedi=0.18.1=py310h06a4308_1
jpeg=9e=h5eee18b_1
json-tricks=3.17.1=pypi_0
jupyter_client=8.1.0=py310h06a4308_0
jupyter_core=5.3.0=py310h06a4308_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libidn2=2.3.4=h5eee18b_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.3=pypi_0
markdown-it-py=3.0.0=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=py310h06a4308_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.6=py310h1128e8f_1
mkl_random=1.2.2=py310h1128e8f_1
mmcv=1.6.1=pypi_0
mmpose=0.28.1=dev_0
model-index=0.1.11=pypi_0
munkres=1.1.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py310h06a4308_0
nettle=3.7.3=hbbd107a_1
numpy=1.25.0=py310h5f9d8c6_0
numpy-base=1.25.0=py310hb5e798b_0
opencv-python=4.7.0.72=pypi_0
opendatalab=0.0.9=pypi_0
openh264=2.1.1=h4ff587b_0
openmim=0.3.9=pypi_0
openssl=3.0.9=h7f8727e_0
ordered-set=4.1.0=pypi_0
packaging=23.1=pypi_0
pandas=2.0.3=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py310h6a678d5_0
pip=23.1.2=py310h06a4308_0
platformdirs=3.8.0=pypi_0
prompt-toolkit=3.0.36=py310h06a4308_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.18.0=pypi_0
pygments=2.15.1=py310h06a4308_1
pyopenssl=23.0.0=py310h06a4308_0
pyparsing=3.1.0=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.11=h955ad1f_3
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.12.1=py3.10_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.1.0=py310h6a678d5_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
rich=13.4.2=pypi_0
scipy=1.11.1=pypi_0
setuptools=67.8.0=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torchaudio=0.12.1=py310_cu113
torchvision=0.13.1=py310_cu113
tornado=6.2=py310h5eee18b_0
tqdm=4.65.0=pypi_0
traitlets=5.7.1=py310h06a4308_0
typing_extensions=4.6.3=py310h06a4308_0
tzdata=2023.3=pypi_0
urllib3=2.0.3=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.38.4=py310h06a4308_0
x264=1!157.20191217=h7b6447c_0
xtcocotools=1.13=pypi_0
xz=5.4.2=h5eee18b_0
yapf=0.40.1=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
